Update Black settings and max_row in generations endpoint
In the .idea/misc.xml file, settings were added for Black, which is Python code formatter. The Python SDK was specified to be Python 3.9 (CurationFlask). This change sets a consistent formatting rule for the project.

Furthermore, in the file endpoints/generations.py, the max_rows value was increased from 100 to 1000. This adjustment allows the generations function to process a larger amount of data.

Also, the logging statement was removed and an option was added to submit_image function to execute without_web